requirements.txt
python-telegram-bot==20.7
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
yfinance==0.2.18
ta==0.10.2
MetaTrader5==5.0.45
schedule==1.2.0
apscheduler==3.10.4
requests==2.31.0
colorama==0.4.6
tqdm==4.66.1
python-dotenv==1.0.0
# Optional/Model related
tensorflow==2.14.0
stable-baselines3==2.1.0
gym==0.26.2
optuna==3.4.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2
peewee==3.17.0
flask==3.0.0
flask-sqlalchemy==3.1.1
websockets==12.0